mytest/requirements.txt
django-annoying
django-bootstrap-toolkit
